added SignIn component and fixed some CSS
