+++ src/App.css
@@ -384,6 +384,8 @@ nav {
 
 .calendarPage-div button {
   border-radius: 8px;
+  padding-left: 20px;
+  padding-right: 20px;
 }
 
 .scheduleBeer {
